requirements.txt
asgiref==3.7.0
chardet==5.1.0
crispy-bootstrap5==2024.2
diff-match-patch==20230430
distlib==0.3.9
Django==5.0
django-admin-sortable2==2.2.3
django-crispy-forms==2.2
django-excel==0.0.10
django-import-export==4.1.1
django-mathfilters==1.0.0
et-xmlfile==1.1.0
filelock==3.16.1
lml==0.1.0
mod_wsgi==5.0.1
mysqlclient==2.2.5
openpyxl==3.1.2
pillow==10.4.0
platformdirs==4.3.6
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pytz==2021.3
sqlparse==0.4.2
tablib==3.5.0
texttable==1.6.7
tzdata==2024.1
virtualenv==20.26.6
